Add namespace, simplify Project and DRY project
